requirements.txt
google
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
enum34==1.1.6
futures==3.2.0
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-vision==0.33.0
googleapis-common-protos==1.5.3
grpcio==1.14.2
idna==2.7
numpy==1.15.1
opencv-python==3.4.2.17
Pillow==5.2.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytesseract==0.2.4
pytz==2018.5
requests==2.19.1
rsa==3.4.2
six==1.11.0
urllib3==1.23
